+++ create/index.html
@@ -381,7 +381,6 @@
         </div>
       </div>
     </footer>
-    <script src="./blocks/signup/__input/signup__input.js"></script>
-    <script src="./blocks/signup/__submit/signup__submit.js"></script>
+    <script src="./script.js"></script>
   </body>
 </html>
