--- pushkin/blocks/pushkin.js
@@ -1,1 +1,0 @@
-import "../../blocks/common.blocks/modal/modal.js";
